move raw_data_viewer to its own file
